requirements.txt
numpy
h5py
python-chess